Enable SETENV to be set for sudo -E
Some tools (e.g. Phabricator's diffusion) make use of `sudo -E`. For
this to work, SETENV needs to be set in the sudoers file. This change
enables the setting of this, via toggle a boolean.

Added test cases for this functionality, and cleaned up a couple of other
failing tests at the same time.