Improve visudo path resolution.
Added a new `visudo_path` attribute to the sudo resource to allow users
to specify the installed path of visudo.

In addition, the setting for the PATH environment variable on SLES and
Solaris does not include the sbin path in which the visudo command is
installed. This ensures the correct path value is inserted into the
PATH environment variable automatically before executing the command.

Fixes #116

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>