Fix non-existent method
Chef runs using the latest release complain with and fail on

```
NoMethodError: undefined method `starts_with' for "ALL":String
```

Presumably because it should be `start_with?`.  I haven't gone the full nine yards to check if this is a complete fix, sorry - just trying to get the ball rolling for a patch.